# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auth",
    "description": "Aicacia Auth API provides authentication services for applications.",
    "contact": {
      "name": "nathanfaucett@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "https://spdx.org/licenses/MIT.html"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/oauth2/{provider}": {
      "post": {
        "tags": [
          "oauth2"
        ],
        "operationId": "create_oauth2_url",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "description": "OAuth2 provider",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "google"
          },
          {
            "name": "register",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          }
        },
        "security": [
          {
            "TenantUUID": []
          }
        ]
      }
    },
    "/oauth2/{provider}/callback": {
      "get": {
        "tags": [
          "oauth2"
        ],
        "operationId": "oauth2_callback",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "description": "OAuth2 provider",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "google"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "302": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          }
        }
      }
    },
...
    "/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "tenant"
        ],
        "operationId": "get_tenant_by_id",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "show_private_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "application_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "tags": [
          "tenant"
        ],
        "operationId": "update_tenant",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "application_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "delete": {
        "tags": [
          "tenant"
        ],
        "operationId": "delete_tenant",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "application_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
    "/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_user_by_id",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "application_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pagination"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashMap"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Algorithm": {
        "type": "string",
        "description": "jsonwebtoken::Algorithm",
        "enum": [
          "HS256",
          "HS384",
          "HS512",
          "ES256",
          "ES384",
          "RS256",
          "RS384",
          "RS512",
          "PS256",
          "PS384",
          "PS512",
          "EdDSA"
        ]
      },
      "Application": {
        "type": "object",
        "required": [
          "id",
          "name",
          "updated_at",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorMessage": {
        "type": "object",
        "required": [
          "code",
          "parameters"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "parameters": {
            "type": "object",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "HashMap": {
        "type": "object",
        "additionalProperties": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/ErrorMessage"
          }
        },
        "propertyNames": {
          "type": "string"
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "has_more",
          "items"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Application"
            }
          }
        }
      },
...
      "Tenant": {
        "type": "object",
        "required": [
          "id",
          "client_id",
          "issuer",
          "algorithm",
          "expires_in_seconds",
          "refresh_expires_in_seconds",
          "oauth2_providers",
          "updated_at",
          "created_at"
        ],
        "properties": {
          "algorithm": {
            "type": "string"
          },
          "audience": {
            "type": [
              "string",
              "null"
            ]
          },
          "client_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "expires_in_seconds": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "issuer": {
            "type": "string"
          },
          "oauth2_providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TenantOAuth2Provider"
            }
          },
          "private_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "public_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "refresh_expires_in_seconds": {
            "type": "integer",
            "format": "int64"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateTenant": {
        "type": "object",
        "properties": {
          "algorithm": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Algorithm"
              }
            ]
          },
          "audience": {
            "type": [
              "string",
              "null"
            ],
            "example": "example.com"
          },
          "client_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "expires_in_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "example": "86400"
          },
          "issuer": {
            "type": [
              "string",
              "null"
            ],
            "example": "Example"
          },
          "private_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "public_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "refresh_expires_in_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "example": "604800"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "TenantUUID": {
        "type": "apiKey",
        "in": "header",
        "name": "Tenant-ID"
      }
    }
  },
...
}